refactor: actions - getVisibleActions cycle 에러 해결
