+++ src/shared/lib/date/formatDate.ts
@@ -1,19 +1,29 @@
-export function getFormattedDate(dateStr: string): string {
+function formatTwoDigits(num: number): string {
+  return num < 10 ? `0${num}` : `${num}`;
+}
+
+export function getFormattedDate(dateStr: string, type: string = ''): string {
   if (!dateStr) return '';
 
   try {
     const date = new Date(dateStr);
-    const month = date.getMonth() + 1;
-    const day = date.getDate();
+    const year = date.getFullYear();
+    const month = formatTwoDigits(date.getMonth() + 1);
+    const day = formatTwoDigits(date.getDate());
 
-    return `${month}월 ${day}일`;
+    switch (type) {
+      case 'dot':
+        return `${year}.${month}.${day}`;
+      default:
+        return `${month}월 ${day}일`;
+    }
   } catch (error) {
     console.error('날짜 형식 변환 오류:', error);
     return '';
   }
 }
 
-export function getFormattedTime(dateStr: string, variant: string = ''): string {
+export function getFormattedTime(dateStr: string, type: string = ''): string {
   if (!dateStr) return '';
 
   try {
@@ -21,12 +31,15 @@ export function getFormattedTime(dateStr: string, variant: string = ''): string 
     const hours = date.getHours();
     const minutes = date.getMinutes();
 
-    const formattedHours = hours < 10 ? `0${hours}` : hours;
-    const formattedMinutes = minutes < 10 ? `0${minutes}` : minutes;
+    const formattedHours = formatTwoDigits(hours);
+    const formattedMinutes = formatTwoDigits(minutes);
 
-    return variant === 'withText'
-      ? `${formattedHours}시 ${formattedMinutes}분`
-      : `${formattedHours}:${formattedMinutes}`;
+    switch (type) {
+      case 'withText':
+        return `${formattedHours}시 ${formattedMinutes}분`;
+      default:
+        return `${formattedHours}:${formattedMinutes}`;
+    }
   } catch (error) {
     console.error('시간 형식 변환 오류:', error);
     return '';
